Add support for custom dependency keywords
Although `compile`, `testCompile`, `runtime` and `testRuntime` are the “regular” dependency block items (or whatever the appropriate term would be), one could in theory add new ones along the road.

For example, `wearApp` and various adaptions of it seems to be used in the Android context, so I figured it would be “easier” to approach the issue this way.